confd: load test-config when in test mode
To enable test mode on a device, a file named /mnt/aux/test-mode
must be manually created.

Although this is uncommon, there are some test scenarios where
retaining the original factory configuration is still necessary.
This factory configuration should only persist until the next reboot.
To facilitate this, we are introducing an additional flag (file) called
/mnt/aux/test-override-startup.